T: Split producers test into multiple files
